\name{print.bgtCI}
\alias{print.bgtCI}

\title{ Print objects of class "bgtCI" }

\description{ Print objects of class "bgtCI" }
\usage{
\method{print}{bgtCI}(x, \dots)
}

\arguments{
  \item{x}{ an object of class "bgtCI" as can be created by calling \code{\link{bgtCI}}}
  \item{\dots}{currently not used}
}

\value{
A print out.
}
\author{  Frank Schaarschmidt, Bioinformatics Unit, University of Hannover, Germany }

\keyword{ print }