\name{print.bgtDesign}
\alias{print.nDesign}
\alias{print.sDesign}

\title{Print Functions for nDesign and sDesign}

\description{
Print function for results of functions \code{\link{nDesign}}, \code{\link{nDesign}}.
}
\usage{
\method{print}{nDesign}(x, \dots)
\method{print}{sDesign}(x, \dots)
}

\arguments{
  \item{x}{ An object of class \kbd{"nDesign"} or \kbd{"sDesign"} }
  \item{\dots}{ currently only \code{digits} is passed to \code{signif} for appropriate rounding }
}

\keyword{ print }