\name{print.bgtvs}
\alias{print.bgtvs}

\title{ Print function for bgtvs }
\description{
  Print function for objects of class "bgtvs".
}
\usage{
print.bgtvs(x, ...)
}
\arguments{
  \item{x}{ An object of class "bgtvs" as can be created by \code{\link{bgtvs}} }
  \item{\dots}{ further arguments to be passed to \code{print} }
}
\value{a print out.
}

\author{ Frank Schaarschmidt }

\keyword{ print }