\name{print.binDesign}
\alias{print.binDesign}

\title{ Print objects of class binDesign }

\description{ 
Print objects of class binDesign
}
\usage{
\method{print}{binDesign}(x, \dots)
}
\arguments{
  \item{x}{ an object of class "binDesign"}
  \item{\dots}{currently not used}
 }
}
\keyword{ print }