\name{print.gt}
\Rdversion{1.1}
\alias{print.gt}
\title{
Print methods for objects of classes "gt" and "gt.mp"
}
\description{
Print methods for objects of classes \kbd{"gt"} and \kbd{"gt.mp"}
}
\usage{
\method{print}{gt}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
objects of class \kbd{"gt"} and \kbd{"gt.mp"}, respectively
}
  \item{digits}{
the digits for rounding
}
  \item{\dots}{
currently not used
}
}

\value{
A print out
}

\keyword{print}
