\name{print.nDesign}
\alias{print.nDesign}

\title{Print for results of nDesign}

\description{
Print function for results of nDesign
}
\usage{
\method{print}{nDesign}(x, \dots)
}

\arguments{
  \item{x}{ An object of class "nDesign" }
  \item{\dots}{ currently not used }
}

\author{ Frank Schaarschmidt, Bioinformatics Unit, University of Hannover, Germany}


}
\keyword{ print }