\name{print.sDesign}
\alias{print.sDesign}

\title{ Print function for sDesign }
\description{
Print function for objects of class sDesign
 }
\usage{

\method{print}{sDesign}(x, \dots)
}

\arguments{
  \item{x}{ an object of class "sDesign" }
  \item{\dots}{ currently only \code{digits} is passed to \code{signif} for appropriate rounding}
  }

\value{ A print out
}

\keyword{ print }