\name{print.bgtvs}
\alias{print.bgtvs}

\title{Print Function for bgtvs}
\description{
  Print function for objects of class \kbd{"bgtvs"}.
}
\usage{

\method{print}{bgtvs}(x, ...)
}
\arguments{
  \item{x}{ An object of class "bgtvs" as can be created by \code{\link{bgtvs}} }
  \item{\dots}{ further arguments to be passed to \code{print} }
}
\value{A print out.
}

\keyword{ print }