\name{print.binDesign}
\alias{print.binDesign}

\title{ Print Function for binDesign }

\description{ 
Print objects of class \kbd{"binDesign"}
}
\usage{
\method{print}{binDesign}(x, \dots)
}
\arguments{
  \item{x}{ an object of class "binDesign"}
  \item{\dots}{currently only \code{digits} is passed to \code{signif} for appropriate rounding}
 }

\keyword{ print }