\name{print.sDesign}
\alias{print.sDesign}

\title{ Print function for sDesign }
\description{
Print function for objects of class sDesign
 }
\usage{

\method{print}{sDesign}(x, \dots)
}

\arguments{
  \item{x}{ an object of class "sDesign" }
  \item{\dots}{ currently not used }
  }

\value{ A print out
}

\author{ Frank Schaarschmidt, Bioinformatics Unit, University of Hannover, Germany}
}
\keyword{ print }